# Storage: reduce memory allocations when merging series sets